# strip tuple wrapper to speed up centroid calc